extracted/miniapp-main/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pyparsing==3.2.3
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
python-multipart>=0.0.9
requests>=2.31.0
google-generativeai>=0.8.3
google-genai
google-ai-generativelanguage
google-api-core>=2.11.0
google-auth==2.32.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.151.0
openai>=1.0.0
anthropic>=0.7.0
httpx>=0.24.0
Pillow>=9.0.0
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
passlib>=1.7.4
tzdata>=2024.2
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0
python-jose>=3.3.0
pandas>=2.2.0
numpy>=1.26.0
jq>=1.6.0
typer>=0.9.0
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.13.0
aiohttp>=3.8.0
litellm>=1.0.0
stripe>=4.0.0
python-telegram-bot>=20.8
cachetools>=5.3.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0
opencv-python>=4.8.0
tqdm>=4.65.0
googleapis-common-protos>=1.62.0
protobuf>=4.25.0
grpcio>=1.60.0
proto-plus>=1.22.0
httplib2>=0.22.0
uritemplate>=4.2.0
distro>=1.8.0
httpcore>=1.0.0
reportlab>=3.6.0
